Random_Racoon/requirements.txt
python-dotenv
webbrowser
openai
google-generativeai